services/payment/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
redis==5.0.1
pydantic==2.5.2
pydantic-settings==2.1.0
stripe==7.14.0
httpx==0.25.2
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
prometheus-fastapi-instrumentator
